// projects/mf-components/src/utils/models/global.model.ts
export class GeneralSchemaFormBuilder {
  fields: FormSchemaField[];
  schemeColumns: FormSchemaLayoutColumn[];
  generalSetting: FormSchemaGeneralSettings;
  variables: string[]
  fetching?: FormSchemaFetching;
  constructor(item?: GeneralSchemaFormBuilder) {
      this.fields = item?.fields ?? [];
      this.schemeColumns = item?.schemeColumns ?? [];
      this.generalSetting = item?.generalSetting ?? new FormSchemaGeneralSettings();
      this.variables = item?.variables ?? [];
      this.fetching = item?.fetching ?? {};
  }
}

export class FormSchemaLayout {
  key: string;
  numberColumns?: number;
  schemaColums: Array<string>;

  constructor(item?: FormSchemaLayout) {
      this.key = item?.key ?? ''
      this.numberColumns = item?.numberColumns ?? 0
      this.schemaColums = item?.schemaColums ?? [];
  }
}

export class FormSchemaLayoutColumn {
  key: string;
  numberColumns: number;
  schemaLayout: Array<FormSchemaLayout>;
  constructor(item?: FormSchemaLayoutColumn) {
      this.key = item?.key ?? ''
      this.numberColumns = item?.numberColumns ?? 0;
      this.schemaLayout = item?.schemaLayout ?? [];
  }
}

export interface FormSchemaPropertyStyle {
  position?: ElementPosition;
  dimensions?: Dimension;
  style?: MobileformsCSSStyleProperties;
}

export class FormSchemaField {
  key: string;
  isArray?: boolean;
  type: string;
  label?: string;
  title?: string;
  value?: string;
  mask?: string;
  specialCharacters?: Array<string>;
  patterns?: IConfig['patterns'];
  required?: boolean;
  disabled?: boolean;
  placeholder?: string;
  options?: Array<SelectionModel>;
  multiline?: boolean;
  eventType?: string;

  /**
   * Table properties
   */
  header?: TableHeader<any>[];
  records?: Array<any>;

  /**
   * Dropdown Element
   */
  optionsVariable?: string;

  /**
   * Button Element
   */
  disableWithControls?: boolean;
  /**
   * For no control elements
   */
  typeReference?: string;
  /**
   * Experimental
   */
  propertiesStyle: FormSchemaPropertyStyle;
  data?: DataFormSchemaField;
  fetching?: FormSchemaFetching

  constructor(item?: FormSchemaField) {
    this.key = item?.key ?? '';
    this.isArray = item?.isArray ?? false;
    this.type = item?.type ?? '';
    this.label = item?.label ?? '';
    this.title = item?.title ?? '';
    this.value = item?.value ?? '';
    this.mask = item?.mask ?? '';
    this.specialCharacters = item?.specialCharacters ?? [];
    this.patterns = item?.patterns ?? {};
    this.required = item?.required ?? false;
    this.disabled = item?.disabled ?? false;
    this.placeholder = item?.placeholder ?? '';
    this.options = item?.options ?? [];
    this.multiline = item?.multiline ?? false;
    this.eventType = item?.eventType ?? '';
    this.propertiesStyle = item?.propertiesStyle ?? {};
    this.typeReference = item?.typeReference ?? '';
    this.header = item?.header ?? [];
    this.fetching = item?.fetching ?? {};
    this.optionsVariable = item?.optionsVariable ?? '';
  }
}

export interface FormSchemaFetching {
  typEvent?: EventTypeElement | null;
  urlFetching?: string;
  variable?: string;
  method?: MethodFetching;
  keys?: {[key: string]: boolean}
}
// ...
export class FormSchemaGeneralSettings {
  endpointUrlBase: string;
  token: string;
  typeForm?: string;

  constructor(item?: FormSchemaGeneralSettings) {
    this.endpointUrlBase = item?.endpointUrlBase ?? '';
    this.token = item?.token ?? '';
    this.typeForm = item?.typeForm ?? '';
  }
}
// ...
export class DefaultKeys {
  _id: string;
  active: number;
  createdAt: number;
  updatedAt: number;

  constructor(item?: DefaultKeys) {
      this._id = item?._id ?? '';
      this.active = item?.active ?? 0;
      this.createdAt = item?.createdAt ?? 0;
      this.updatedAt = item?.updatedAt ?? 0;
  }
}

export class Form extends DefaultKeys {
  name: string;
  description: string;
  settings: GeneralSchemaFormBuilder;

  constructor(item?: Form) {
      super(item);
      this.name = item?.name ?? ''
      this.description = item?.description ?? ''
      this.settings = item?.settings ?? new GeneralSchemaFormBuilder();
  }
}


// Cache loader
export interface SocketReceivers<T> {
  fullDocument: T,
  operationType: 'insert' | 'delete' |'replace' |'update' |'drop' |'rename' |'dropDatabase' | 'invalidate';
  documentKey: {_id: string}
  updateDescription: UpdateDescription
}

export interface UpdateDescription {
  removedFields: Array<any>,
  truncatedArrays: Array<any>,
  updatedFields: {[key: string]: any}
}